Add type scope implementation that I'm finally happy with
Type definitions go the route of "builtin vs user-defined" similar to
functions. Builtin types, unlike builtin functions, can't really be
treated in the same way as their user-defined counterpart. They are
generally more foundational, and although are checked at runtime, are
going to be more of a finite set than all builtin functions. As a
result, it's not such an anomaly to have a special bytecode instruction
for checking a known built-in type. On the other hand, having a bytecode
instruction to correspond to every single builtin function is generally
insanity.

That's my rationalization. I'm probably going to gut the system another
two or three times before I'm *actually* happy with it. Without further
ado, here is what was changed:

- vm::Ty is now either a Ty::User(Symbol) or Ty::Builtin(BuiltinTy)
- vm::BuiltinTy is just a finite set (enum) of known built-in types
- ir::TyExpr has a Builtin variant for known builtins
- Add ir::compile::TyScope for tracking types
- Builtin functions now hold a BuiltinTy instead of a Ty
- Misc cleanup/squashing of warnings

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>