Update vm::Ty::Builtin(...) to have a symbol
I think this was originally set up because I wasn't able to search the
type scope for a while when adding return types a long while ago. This
was a kind of annoying thing to have, but now I can feel justified in
throwing types around the VM because I'll always have a symbol.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>