Initial backend rewrite commit.
This is pretty large and the changes are numerous. The biggest ones are:

* ir, compiler, vm: New IR and VM layout, with some API ideas borrowed from the previous
  implementation.
* compiler: Starting to use "transform" pattern, where individual states
  for specific "transforms" are retained in the compile state, one at a
  time
* compiler: Starting to add support for better error messages with
  fallible operations
* syntax: Identifiers that begin with an upper case character are
  denoted to be types. Else they are functions

Overall, things have been rewritten to work better together. The
compiler and VM still need to be finished but we are getting there. The
structure exists, and hopefully it shouldn't be too difficult to
implement past this point.

Signed-off-by: Alek Ratzloff <alekratz@gmail.com>